Large refactor, changed system for retrieving annotations. Moved algorithms to separate files. Changed changeAlgo and selectedFunction. Moved some globals into an AnimationController.
