+++ index.html
@@ -241,7 +241,7 @@
                                 <p class="text">A* is <strong>more efficient</strong> than Djikstra's and guarantees the shortest path between two points.</p>
                             </div>
                             <div class="card-footer">
-                                <button onclick="window.location.href='pathfinding/pathfind.html?func=a*'" class="btn text menu-button btn-navajoWhite justify-content-end" ><strong>Go</strong></button>
+                                <button onclick="window.location.href='pathfinding/pathfind.html?func=astar'" class="btn text menu-button btn-navajoWhite justify-content-end" ><strong>Go</strong></button>
                             </div>
                         </div>
                     </div>
