+++ js/AnimationController.js
@@ -1,0 +1,7 @@
+export class AnimationController {
+    constructor(params) {
+    }
+    static animationSpeed = 1
+    static speeds = [1, 2, 4]
+    static currentAnim = undefined
+}
