+++ js/PathfindingAnimations.js
@@ -1,0 +1,92 @@
+import { Action } from "./Action.js"
+import { AnimationController } from "./AnimationController.js"
+import anime from "./anime.es.js"
+
+class FinalPath extends Action {
+    constructor(targets, line=2) {
+        super(targets, line)
+        FinalPath.duration = 1000
+    }
+
+    get duration() {
+        return FinalPath.duration / AnimationController.animationSpeed
+    }
+
+    get annotation() {
+        return `The algorithm found the end. To find the path, it backtracks to the parent of the end node, then continues backtracking until the start.`
+    }
+
+    get animation() {
+        return {
+            targets: this.targets,
+            delay: anime.stagger(50),
+            duration: 500,
+            backgroundColor: '#FEDC97',
+        }
+    }
+
+    async Animate() {
+        AnimationController.currentAnim = anime(this.animation)
+        return AnimationController.currentAnim.finished
+    }
+}
+
+class SearchedPath extends Action {
+    constructor(targets, line=1) {
+        super(targets, line)
+        SearchedPath.duration = 500
+    }
+
+    get duration() {
+        return SearchedPath.duration / AnimationController.animationSpeed
+    }
+
+    get annotation() {
+        return `This was the best available path, but it did not reach the end. The algorithm will add any new neighbors and pick the next best path.`
+    }
+
+    get animation() {
+        return {
+            targets: this.targets,
+            backgroundColor: [
+                { value: "#F26419", duration: 0 },
+                { value: "#28666E", delay: this.duration/AnimationController.animationSpeed, duration: 1 } //Small wait, then zap the whole line purple
+            ]
+        }
+    }
+
+    async Animate() {
+        AnimationController.currentAnim = anime(this.animation)
+        return AnimationController.currentAnim.finished
+    }
+}
+
+class NewChildren extends Action {
+    constructor(targets, line=3) {
+        super(targets, line)
+        NewChildren.duration = 100
+    }
+
+    get duration() {
+        return NewChildren.duration / AnimationController.animationSpeed
+    }
+
+    get annotation() {
+        return `Adding any new neighbors to the list of potential paths.`
+    }
+
+    get animation() {
+        return {
+            targets: this.targets,
+            duration: this.duration,
+            backgroundColor: "#696464"
+        }
+    }
+
+    async Animate() {
+        AnimationController.currentAnim = anime(this.animation)
+        return AnimationController.currentAnim.finished
+    }
+}
+
+export { FinalPath, SearchedPath, NewChildren }
