+++ pathfinding/pathfind.html
@@ -196,7 +196,7 @@
                 <div class="btn-group dropup col-auto gx-0">
                     <button type="button" class="btn btn-menu dropdown-toggle d-none d-sm-block d-sm-none d-md-block text" data-bs-toggle="dropdown" aria-expanded="false">Pathfinding</button>
                     <ul class="dropdown-menu">
-                        <li><a class="dropdown-item text" id="a*">A* Pathfinding</a></li>
+                        <li><a class="dropdown-item text" id="astar">A* Pathfinding</a></li>
                         <li><a class="dropdown-item text" id="djikstra">Djikstra's</a></li>
                         <!-- <li><a class="dropdown-item" onclick="changeAlgo('bfs')">BFS</a></li> -->
                     </ul>
